Modify git.py file

+++ git.py
@@ -58,7 +58,10 @@ def unset_proxy(pc_username):
         if pc_username == 'root' :
             file_name = '/root/.gitconfig'
             file_temp = open(file_name + '_temp', 'w+')
-            file_ob = open(file_name, 'w+')
+            if os.path.exists(file_name):
+                file_ob = open(file_name, 'r')
+            else:
+                file_ob = open(file_name, 'w+')
             for line in file_ob.readlines() :
                 if 'http' in line or 'https' in line or 'proxy' in line:
                     continue    
@@ -72,7 +75,10 @@ def unset_proxy(pc_username):
         else:
             file_name = '/home/'+ pc_username + '/.gitconfig'
             file_temp = open(file_name + '_temp', 'w+')
-            file_ob = open(file_name, 'w+')
+            if os.path.exists(file_name):
+                file_ob = open(file_name, 'r')
+            else:
+                file_ob = open(file_name, 'w+')
             for line in file_ob.readlines():
                 if 'http' in line or 'https' in line or 'proxy' in line:
                     continue
